# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Sign in"
                ],
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "center@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "center"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register/student": {
            "post": {
                "tags": [
                    "Sign up"
                ],
                "description": "Sign up as student",
                "operationId": "studentRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "student123@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "student"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "First name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Last name"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "+374 98-066-083"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register/trainer": {
            "post": {
                "tags": [
                    "Sign up"
                ],
                "description": "Sign up as trainer",
                "operationId": "trainertRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "traine@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "traine"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "First name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Last name"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "+374 98-066-083"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}